On object creation defer widget text generation
When a text widget is created, it's text needs to be generated quickly in order
to avoid a "popup" glitch. It used to be done by calling the update method
irecly in the initializor. This method implies that awesome will fail to load
if any update function throws an error.

Defer execution by sending a signal emulate a timer timeout.